requirements.txt
# Core Framework
fastapi
uvicorn
starlette
python-multipart
python-dotenv
requests

# Data Validation
pydantic
pydantic[email]
email-validator

# Database
pymongo

# AI & LLM
groq

# PDF Processing
pdfplumber
PyPDF2
PyMuPDF

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pyjwt
cryptography

# Webhooks (Clerk)
svix